Part service

+++ part-service/src/main/java/com/oem/evpart/dto/response/PartAllocationResponse.java
@@ -14,7 +14,7 @@ public class PartAllocationResponse {
     private Long inventoryId;
     private String partName;
     private Long serviceCenterId;
-    private Integer allocatedQty;
+    private Long allocatedQty;
     private LocalDateTime allocatedDate;
 }
