+++ part-service/src/main/java/com/oem/evpart/models/PartAllocation.java
@@ -25,7 +25,7 @@ public class PartAllocation {
     private Long serviceCenterId; // FK từ service khác (User-Service)
 
     @Column(name = "allocated_qty", nullable = false)
-    private Integer allocatedQty;
+    private Long allocatedQty;
 
     @Column(name = "allocated_date", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
     private LocalDateTime allocatedDate;
